Modify the user controller by using helper functions
